+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed error while passing date parameter in execute
+
 ## 2.12.2 ##
 * Fix error of check retriable error for idempotent operations (error exist since 2.12.1)