+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## 2.12.3 ##
 * Add six package to requirements
 * Fixed error while passing date parameter in execute